refactor: restructure the project
This commit introduces a significant restructuring of the project to improve organization and maintainability.
- Moved `cookies-consent` component to `src/app/components/`.
- Moved `about-us` component to `src/app/pages/about-us/`.
- Moved `main` component to `src/app/pages/home/`.
- Moved `research-projects` component to `src/app/pages/research-projects/`.
- Replaced the `pictures` component with the new `gallery` component.
- Modified `app.routes.ts` to reflect the new structure and include the gallery component.
- Updated `footer.component.html` and `header.component.html` to align with the new structure and include links to the new gallery page.
- Adjusted imports and references in `app.component.ts` to accommodate the refactored structure.

Rationale:
- The restructuring organizes components into `components` and `pages` directories, making the project easier to navigate and scale.
- The new gallery component replaces the pictures component, providing a more general usecase.

This refactor lays the groundwork for future development and improves the overall maintainability of the codebase.
