--- src/app/about-us/about-us.component.scss
@@ -1,212 +1,0 @@
-.about {
-    text-align: center;
-    padding: 20px;
-    background-color: #1864ab;
-    color: white;
-}
-
-.social-links {
-    display: flex;
-    gap: 10px;
-    justify-content: center;
-    font-size: 24px;
-    color: #555;
-}
-
-.social-links a {
-    color: inherit;
-    text-decoration: none;
-}
-
-.social-links a:hover {
-    color: #000;
-}
-
-.content-container {
-    display: flex;
-    justify-content: space-between;
-    padding: 20px;
-}
-
-.professors-section,
-.exec-team-section {
-    padding: 20px;
-    background: white;
-    margin-bottom: 20px;
-}
-
-.professor-cards,
-.exec-cards {
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: space-around;
-    gap: 20px;
-}
-
-.professor-card,
-.exec-card {
-    text-align: center;
-    padding: 10px;
-    background: white;
-    border-radius: 8px;
-    box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
-    width: 200px;
-}
-
-.professor-card img,
-.exec-card img {
-    width: 150px;
-    height: 150px;
-    object-fit: cover;
-    border-radius: 50%;
-}
-
-.slider-container {
-    position: relative;
-    width: 100%;
-    max-width: 800px;
-    margin: 20px auto;
-    overflow: hidden;
-}
-
-.slider-wrapper {
-    width: 100%;
-    height: 400px;
-    overflow: hidden;
-}
-
-.slider {
-    display: flex;
-    transition: transform 0.5s ease;
-    height: 100%;
-}
-
-.slide {
-    min-width: 100%;
-    height: 100%;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    padding: 20px;
-    box-sizing: border-box;
-    background-color: #fff;
-    border-radius: 8px;
-    box-shadow: 0 2px 10px rgba(0, 0, 0, 0.1);
-}
-
-.slide img {
-    width: 200px;
-    height: 200px;
-    object-fit: cover;
-    border-radius: 50%;
-    margin-bottom: 15px;
-}
-
-.slide-content {
-    text-align: center;
-}
-
-.slider-controls {
-    display: flex;
-    justify-content: center;
-    margin-top: 15px;
-}
-
-.slider-btn {
-    background-color: #333;
-    color: white;
-    border: none;
-    width: 40px;
-    height: 40px;
-    border-radius: 50%;
-    margin: 0 5px;
-    cursor: pointer;
-    font-size: 18px;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-}
-
-.slider-btn:hover {
-    background-color: #555;
-}
-
-.slider-dots {
-    display: flex;
-    justify-content: center;
-    margin-top: 10px;
-}
-
-.dot {
-    width: 10px;
-    height: 10px;
-    border-radius: 50%;
-    background-color: #ccc;
-    margin: 0 5px;
-    cursor: pointer;
-}
-
-.dot.active {
-    background-color: #333;
-}
-
-.main-content {
-    display: flex;
-    flex-direction: column;
-}
-
-.upper-section {
-    display: flex;
-    flex-wrap: wrap;
-}
-
-.slider-section,
-.calendar-section,
-.carousel-section {
-    flex: 1;
-    min-width: 300px;
-    padding: 20px;
-}
-
-.lower-section {
-    display: flex;
-    flex-direction: column;
-    padding: 0 20px;
-}
-
-.carousel-wrapper {
-    overflow: hidden;
-    position: relative;
-}
-
-.cards {
-    display: flex;
-    transition: transform 0.5s ease;
-    height: 100%;
-}
-
-.card {
-    min-width: 100%;
-    padding: 20px;
-    box-sizing: border-box;
-    text-align: center;
-}
-
-.carousel-controls {
-    display: flex;
-    justify-content: center;
-    margin-top: 15px;
-}
-
-.carousel-controls button {
-    background-color: #333;
-    color: white;
-    border: none;
-    width: 40px;
-    height: 40px;
-    border-radius: 50%;
-    margin: 0 5px;
-    cursor: pointer;
-    font-size: 18px;
-}
