+++ src/app/app.component.ts
@@ -4,7 +4,7 @@ import { isPlatformBrowser } from '@angular/common';
 import { Subscription } from 'rxjs';
 import { filter } from 'rxjs/operators';
 import { gsap } from 'gsap';
-import { CookiesConsentComponent } from './cookies-consent/cookies-consent.component';
+import { CookiesConsentComponent } from './components/cookies-consent/cookies-consent.component';
 
 @Component({
     selector: 'app-root',
