+++ src/app/pages/about-us/(Old)AboutUs.html
@@ -116,7 +116,7 @@
         <nav>
             <a href="javascript:void(0)">Research and Projects</a>
             <a href="javascript:void(0)">Opportunities & Volunteering</a>
-            <a href="javascript:void(0)">Pictures</a>
+            <a href="javascript:void(0)">Gallery</a>
         </nav>
     </header>
