+++ src/app/pages/gallery/gallery.component.html
@@ -1,0 +1,16 @@
+<!-- Gallery Component: Image gallery showcase -->
+
+<section class="about">
+  <h1>Gallery</h1>
+  <p>
+    These images showcase the vibrant spirit of CODERS Club as
+    we engage with our community and foster a culture of innovation.
+    From hands-on coding sessions with local youth to collaborative
+    events with fellow clubs on campus, every moment is an opportunity
+    to inspire and be inspired. Our photos highlight the joy of
+    learning, the power of collaboration, and the impact of
+    community service.
+  </p>
+</section>
+
+<p>PICTURES GO HERE, VIDEOS TOO IF ANY</p>
