+++ src/app/pages/gallery/gallery.component.spec.ts
@@ -1,0 +1,25 @@
+// Gallery spec: Testing component logic and functionality
+
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { GalleryComponent } from './gallery.component';
+
+describe('GalleryComponent', () => {
+  let component: GalleryComponent;
+  let fixture: ComponentFixture<GalleryComponent>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      imports: [GalleryComponent]
+    })
+    .compileComponents();
+
+    fixture = TestBed.createComponent(GalleryComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
