+++ src/app/pages/gallery/gallery.component.ts
@@ -1,0 +1,11 @@
+// Gallery functionality: Image loading and transitions
+
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-gallery',
+  imports: [],
+  templateUrl: './gallery.component.html',
+  styleUrl: './gallery.component.scss'
+})
+export class GalleryComponent {}
