+++ src/app/pages/home/home.component.spec.ts
@@ -2,7 +2,7 @@
 
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 
-import { MainComponent } from './main.component';
+import { MainComponent } from './home.component';
 
 describe('MainComponent', () => {
   let component: MainComponent;
