+++ src/app/pages/home/home.component.ts
@@ -4,11 +4,9 @@ import { Component } from '@angular/core';
 import { RouterModule } from '@angular/router';
 
 @Component({
-  selector: 'app-main',
-  templateUrl: './main.component.html',
-  styleUrl: './main.component.scss',
+  selector: 'app-home',
+  templateUrl: './home.component.html',
+  styleUrl: './home.component.scss',
   imports: [RouterModule]
 })
-export class MainComponent {
-
-}
+export class MainComponent {}
