+++ src/app/pages/research-projects/research-projects.component.scss
@@ -1,0 +1,12 @@
+.research-projects {
+    text-align: center;
+    padding: 20px;
+    h1 {
+        border-bottom: 1px solid #aaa;
+        padding: 10px;
+        display: inline-block;
+    }
+    p {
+        font-style: italic;
+    }
+}
