--- src/app/pictures/pictures.component.html
@@ -1,20 +1,0 @@
-<!-- Pictures component: Image gallery showcase -->
-
-<section class="about">
-    <h1>Pictures</h1>
-    <p>
-        These images showcase the vibrant spirit of CODERS Club as 
-        we engage with our community and foster a culture of innovation. 
-        From hands-on coding sessions with local youth to collaborative 
-        events with fellow clubs on campus, every moment is an opportunity 
-        to inspire and be inspired. Our photos highlight the joy of 
-        learning, the power of collaboration, and the impact of 
-        community service.
-    </p>
-  </section>
-
-<br>
-
-PICTURE SECTION GOES HERE
-
-<br><br>
