--- src/app/pictures/pictures.component.spec.ts
@@ -1,25 +1,0 @@
-// Pictures spec: Testing component logic and functionality
-
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { PicturesComponent } from './pictures.component';
-
-describe('PicturesComponent', () => {
-  let component: PicturesComponent;
-  let fixture: ComponentFixture<PicturesComponent>;
-
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      imports: [PicturesComponent]
-    })
-    .compileComponents();
-
-    fixture = TestBed.createComponent(PicturesComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
