--- src/app/pictures/pictures.component.ts
@@ -1,13 +1,0 @@
-// Pictures functionality: Image loading and transitions
-
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-pictures',
-  imports: [],
-  templateUrl: './pictures.component.html',
-  styleUrl: './pictures.component.scss'
-})
-export class PicturesComponent {
-
-}
