+++ src/app/shared/header/header.component.html
@@ -31,7 +31,7 @@
           <li><a (click)="navigateAndReload('/home')">Home</a></li>
           <li><a (click)="navigateAndReload('/research')">Research & Projects</a></li>
           <li><a (click)="navigateAndReload('/opportunities')">Opportunities</a></li>
-          <li><a (click)="navigateAndReload('/pictures')">Pictures</a></li>
+          <li><a (click)="navigateAndReload('/gallery')">Gallery</a></li>
           <li><a (click)="navigateAndReload('/aboutus')">About Us</a></li>
         </ul>
       </nav>
